Keep virtio block IRQ completion independent of SYS queue state
The Linx rootfs bring-up stalled after the first virtio block read because vm_interrupt loaded a 32-bit MMIO interrupt status into an unsigned long. The generated code then tested status with a queued u#1 mask across a BSTART.SYS fence blocklet, and QEMU's scalar queue reset made the interrupt look inactive even after the device asserted it.

Use the register-width u32 type for the virtio-mmio interrupt status so the generated test stays local to the loaded value and the handler reaches vring_interrupt, virtblk_done, and blk_mq_end_request on the first completed block request.

Constraint: Virtio-mmio interrupt status is a 32-bit device register on the Linx rootfs lane.

Rejected: Paper over the symptom in the driver with inline assembly or volatile mask materialization | the source type was wider than the MMIO register contract and kept the generated code dependent on queue lifetime.

Confidence: high

Scope-risk: narrow

Directive: Do not widen vm_interrupt status without checking Linx disassembly and a rootfs virtio completion trace.

Tested: clean Linx vmlinux build with run_linux_vmlinux_build_clean.sh; rootfs trace reached vring_interrupt, virtblk_done, blk_mq_complete_request, and blk_mq_end_request after the first virtio read.

Not-tested: full BusyBox rootfs shell and SPECINT; the rootfs repro still has a later boot blocker after block completion.

+++ drivers/virtio/virtio_mmio.c
@@ -322,7 +322,7 @@ static irqreturn_t vm_interrupt(int irq, void *opaque)
 {
 	struct virtio_mmio_device *vm_dev = opaque;
 	struct virtqueue *vq;
-	unsigned long status;
+	u32 status;
 	irqreturn_t ret = IRQ_NONE;
 
 	/* Read and acknowledge interrupts */
